requirements.txt
CairoSVG==2.5.1
hjson==3.0.2
matplotlib==3.2.2
numpy==1.19.4
penkit==0.0.6
pyserial==3.5
requests==2.25.1
scipy==1.5.4
svgpathtools==1.4.1
svgwrite==1.4
PyQt-builder==1.8.0
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
svgelements==1.4.7
